* feat(Cargo.toml): add pcu_lib as a library dependency and pcu as a binary * fix(change_frag.rs): rename log_update.rs to change_frag.rs and update imports and references accordingly * feat(pr_title.rs): add PrTitle struct and parse function for parsing PR titles * feat(main.rs): update imports and references to use pcu_lib instead of log_update * feat(lib.rs): add module declarations for change_frag and pr_title and re-export them from the library * test(pr_title.rs): add unit tests for parsing PR titles
