+++ Cargo.toml
@@ -18,3 +18,11 @@ tokio = { version = "1.37.0", features = ["macros", "rt-multi-thread"] }
 
 [dev-dependencies]
 rstest = "0.19.0"
+
+[lib]
+name = "pcu_lib"
+path = "src/lib.rs"
+
+[[bin]]
+name = "pcu"
+path = "src/main.rs"
