+++ src/main.rs
@@ -1,8 +1,6 @@
 use std::env;
 
-use crate::log_update::{ChangeRecord, PrTitle};
-
-mod log_update;
+use pcu_lib::{ChangeFrag, PrTitle};
 
 #[tokio::main]
 async fn main() {
@@ -35,7 +33,7 @@ async fn changelog_update() -> Result<(), octocrab::Error> {
         if let Some(title) = pr.title.as_ref() {
             let pr_title = PrTitle::parse(title);
             println!("PR: {:#?}", pr_title);
-            let change_frag = ChangeRecord::from(&pr_title);
+            let change_frag = ChangeFrag::from(&pr_title);
             println!("Change record: {:#?}", change_frag);
         }
     });
